Add A-Packages increment
