+++ src/main/java/command/DeleteCommand.java
@@ -1,3 +1,9 @@
+package command;
+
+import parser.Storage;
+import task.TaskList;
+import ui.Ui;
+
 public class DeleteCommand extends Command {
     private int index;
