+++ src/main/java/command/DoneCommand.java
@@ -1,3 +1,9 @@
+package command;
+
+import parser.Storage;
+import task.TaskList;
+import ui.Ui;
+
 public class DoneCommand extends Command {
     private int index;
