+++ src/main/java/command/IncorrectCommand.java
@@ -1,3 +1,9 @@
+package command;
+
+import parser.Storage;
+import task.TaskList;
+import ui.Ui;
+
 public class IncorrectCommand extends Command {
     @Override
     public void execute(TaskList tasks, Ui ui, Storage storage) {
