+++ src/main/java/converter/StringDateConverter.java
@@ -1,3 +1,5 @@
+package converter;
+
 import java.text.SimpleDateFormat;
 import java.util.Date;
