+++ src/main/java/task/Deadline.java
@@ -1,3 +1,5 @@
+package task;
+
 import java.util.Date;
 
 public class Deadline extends Task {
