+++ src/main/java/task/Event.java
@@ -1,3 +1,5 @@
+package task;
+
 import java.util.Date;
 
 public class Event extends Task {
