+++ src/main/java/ui/Ui.java
@@ -1,3 +1,5 @@
+package ui;
+
 import java.util.Scanner;
 
 public class Ui {
